+++ NEWS.md
@@ -3,6 +3,7 @@
 * fixed wrong percentage in `pw_report()`
 * fixed iterator being accessed outside loop
 * more reliable datetime parsing with `anytime`
+* datetime parsing on js level if readability.js doesnt find a date 
 
 # paperwizard 0.1.0